add ingestion pipeline for default, pg and mongo
