+++ templates/components/vectordbs/python/mongo/index.py
@@ -2,7 +2,7 @@ import logging
 import os
 
 from llama_index.core.indices import VectorStoreIndex
-from llama_index.vector_stores.mongodb import MongoDBAtlasVectorSearch
+from app.engine.vectordb import get_vector_store
 
 
 logger = logging.getLogger("uvicorn")
@@ -10,11 +10,7 @@ logger = logging.getLogger("uvicorn")
 
 def get_index():
     logger.info("Connecting to index from MongoDB...")
-    store = MongoDBAtlasVectorSearch(
-        db_name=os.environ["MONGODB_DATABASE"],
-        collection_name=os.environ["MONGODB_VECTORS"],
-        index_name=os.environ["MONGODB_VECTOR_INDEX"],
-    )
+    store = get_vector_store()
     index = VectorStoreIndex.from_vector_store(store)
     logger.info("Finished connecting to index from MongoDB.")
     return index
