--- templates/components/vectordbs/python/none/constants.py
@@ -1,1 +1,0 @@
-STORAGE_DIR = "storage"  # directory to cache the generated index
