--- templates/components/vectordbs/python/pg/constants.py
@@ -1,2 +1,0 @@
-PGVECTOR_SCHEMA = "public"
-PGVECTOR_TABLE = "llamaindex_embedding"
