+++ templates/components/vectordbs/python/pg/index.py
@@ -1,13 +1,13 @@
 import logging
 from llama_index.core.indices.vector_store import VectorStoreIndex
-from app.engine.utils import init_pg_vector_store_from_env
+from app.engine.vectordb import get_vector_store
 
 logger = logging.getLogger("uvicorn")
 
 
 def get_index():
     logger.info("Connecting to index from PGVector...")
-    store = init_pg_vector_store_from_env()
+    store = get_vector_store()
     index = VectorStoreIndex.from_vector_store(store)
     logger.info("Finished connecting to index from PGVector.")
     return index
