+++ templates/types/streaming/fastapi/app/constants.py
@@ -1,0 +1,1 @@
+STORAGE_DIR = "storage"  # directory to cache the document store
